Worker cleanup Activity 구현을 전용 모듈로 분리한다
